requirements.txt
asgiref==3.6.0
autopep8==1.7.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssselect2==0.7.0
distlib==0.3.4
Django==4.1.7
django-appconf==1.0.5
django-compressor==4.3.1
django-filer==2.2.4
django-js-asset==2.0.0
django-libsass==0.9
django-markdowny==0.2.4
django-mdeditor==0.1.20
django-modeltranslation==0.18.9
django-mptt==0.14.0
django-polymorphic==3.1.0
django-rosetta==0.9.9
django-throttle-requests==0.7.0
django_debug_toolbar==3.8.1
easy-thumbnails==2.8.5
filelock==3.7.0
Flask==2.2.3
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
libsass==0.22.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
Pillow==9.5.0
platformdirs==2.5.2
playwright==1.31.1
polib==1.2.0
pycodestyle==2.9.1
pyee==9.0.4
rcssmin==1.1.1
reportlab==3.6.12
requests==2.28.2
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
toml==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
Unidecode==1.1.2
urllib3==1.26.15
virtualenv==20.14.1
webencodings==0.5.1
Werkzeug==2.2.3
